chat-bot-db/requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dis==1.0.1
distlib==0.3.7
Django==4.2.3
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
dlib==19.24.2
docopt==0.6.2
drf-yasg==1.21.7
fastjsonschema==2.18.1
filelock==3.12.2
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_core==5.3.2
magic-filter==1.0.11
multidict==6.0.4
nbformat==5.9.2
packaging==23.2
pigar==2.1.1
pipreqs==0.4.13
platformdirs==3.9.1
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.14.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.4
sqlparse==0.4.4
telebot==0.0.5
traitlets==5.11.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.24.0
yarg==0.1.9
yarl==1.9.2
